# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Modi Doodh Dairy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Auth operations
  - name: orders
    description: Order operations
  - name: reviews
    description: Review operations
  - name: referral
    description: Referral operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders for current user
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Cannot cancel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{orderId}/return:
    post:
      operationId: returnOrder
      tags: [orders]
      summary: Request a return for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Return requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a product review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "200":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/{productId}:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  /referral/apply:
    post:
      operationId: applyReferral
      tags: [referral]
      summary: Apply referral code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyReferralRequest"
      responses:
        "200":
          description: Referral applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralResponse"
        "400":
          description: Invalid referral code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
      required:
        - name
        - phone
        - password

    LoginRequest:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required:
        - phone
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        referralCode:
          type: string
        referralBalance:
          type: number
      required:
        - id
        - name
        - phone
        - referralCode
        - referralBalance

    OrderItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        price:
          type: number
        totalPrice:
          type: number
      required:
        - productId
        - productName
        - quantity
        - unit
        - price
        - totalPrice

    CreateOrderRequest:
      type: object
      properties:
        userId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        deliveryAddress:
          type: string
        deliveryState:
          type: string
        deliveryCity:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        paymentMethod:
          type: string
          enum: [upi, cod]
        subtotal:
          type: number
        deliveryCharge:
          type: number
        totalAmount:
          type: number
        referralDiscount:
          type: number
      required:
        - customerName
        - customerPhone
        - deliveryAddress
        - deliveryState
        - deliveryCity
        - items
        - paymentMethod
        - subtotal
        - deliveryCharge
        - totalAmount
        - referralDiscount

    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        userId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        deliveryAddress:
          type: string
        deliveryState:
          type: string
        deliveryCity:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        paymentMethod:
          type: string
        subtotal:
          type: number
        deliveryCharge:
          type: number
        totalAmount:
          type: number
        referralDiscount:
          type: number
        status:
          type: string
          enum: [pending, submitted, confirmed, processing, shipped, delivered, cancelled, return_requested]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - customerName
        - customerPhone
        - deliveryAddress
        - deliveryState
        - deliveryCity
        - items
        - paymentMethod
        - subtotal
        - deliveryCharge
        - totalAmount
        - referralDiscount
        - status
        - createdAt
        - updatedAt

    CreateReviewRequest:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        productId:
          type: string
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - productId
        - rating
        - userName

    Review:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        productId:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userName
        - productId
        - rating
        - createdAt

    ApplyReferralRequest:
      type: object
      properties:
        userId:
          type: string
        referralCode:
          type: string
      required:
        - referralCode

    ReferralResponse:
      type: object
      properties:
        success:
          type: boolean
        discount:
          type: number
        message:
          type: string
      required:
        - success
        - discount
        - message